Fix createStoreProvider bug
which was unsubscribing the connectors whenever the state changed. This behavior is not allowed because the app needs to keep all data source connections alive during its execution, for instance, the Subjects contexts weren't being kept. With this new implementation, the state changes is delegated to the oracleSubject and the connections are now being subscribed inside a unchangeable useEffect hook.